+++ fitness-app/src/Components/Footer.jsx
