+++ fitness-app/src/Components/Navbar.jsx
