+++ fitness-app/src/Components/QuoteSection.jsx
