+++ fitness-app/src/pages/Activity.jsx
